Utrition/requirements.txt
certifi==2020.12.5
click==7.1.2
elasticsearch==7.12.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
urllib3==1.26.4
Werkzeug==1.0.1
